requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.39.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pygame==2.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.0
urllib3==1.26.15
webdriver-manager==3.8.5
wsproto==1.2.0
mss